+++ packages/transaction-pay-controller/src/TransactionPayController.ts
@@ -26,7 +26,12 @@ import type {
 import { getStrategyOrder } from './utils/feature-flags';
 import { updateQuotes } from './utils/quotes';
 import { updateSourceAmounts } from './utils/source-amounts';
-import { getTransaction, pollTransactionChanges } from './utils/transaction';
+import { buildCaipAssetType } from './utils/token';
+import {
+  getTransaction,
+  subscribeAssetChanges,
+  subscribeTransactionChanges,
+} from './utils/transaction';
 
 const MESSENGER_EXPOSED_METHODS = [
   'getDelegationTransaction',
@@ -87,10 +92,16 @@ export class TransactionPayController extends BaseController<
       MESSENGER_EXPOSED_METHODS,
     );
 
-    pollTransactionChanges(
+    subscribeTransactionChanges(
       messenger,
       this.#updateTransactionData.bind(this),
       this.#removeTransactionData.bind(this),
+    );
+
+    subscribeAssetChanges(
+      messenger,
+      () => this.state,
+      this.#updateTransactionData.bind(this),
     );
 
     // eslint-disable-next-line no-new
@@ -284,7 +295,10 @@ export class TransactionPayController extends BaseController<
         transactionId,
         this.messenger,
       ) as TransactionMeta;
-      const fiatAsset = deriveFiatAssetForFiatPayment(transaction);
+      const fiatAsset = deriveFiatAssetForFiatPayment(
+        transaction,
+        this.messenger,
+      );
       if (fiatAsset) {
         this.#updateTransactionData(transactionId, (data) => {
           if (data.fiatPayment) {